Change block worker load api's bandwidth to int64
### What changes are proposed in this pull request?

Use bytes instead of MB as the unit for bandwidth

pr-link: Alluxio/alluxio#15990

+++ core/server/worker/src/main/java/alluxio/worker/page/PagedBlockStore.java
@@ -46,7 +46,6 @@ import java.io.IOException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.OptionalInt;
 import java.util.OptionalLong;
 import java.util.Set;
 import java.util.concurrent.CopyOnWriteArrayList;
@@ -186,7 +185,7 @@ public class PagedBlockStore implements BlockStore {
   }
 
   @Override
-  public List<BlockStatus> load(List<Block> fileBlocks, String tag, OptionalInt bandwidth) {
+  public List<BlockStatus> load(List<Block> fileBlocks, String tag, OptionalLong bandwidth) {
     return null;
   }
